+++ libs/contactus/shared/src/lib/components/members-list/members-list.component.ts
@@ -29,7 +29,7 @@ import { listAddRemoveAnimation, SpaceTypeFamily } from '@sneat/core';
 import {
   ScheduleNavService,
   ScheduleNavServiceModule,
-} from '@sneat/mod-schedulus-core';
+} from '@sneat/mod-calendarius-core';
 import { WithSpaceInput } from '@sneat/space-services';
 import { ContactService, ContactusNavService } from '@sneat/contactus-services';
 import {