+++ libs/extensions/calendarius/shared/src/lib/services/calendarius-services.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { HappeningSlotModalService } from '../components/happening-slot-form/happening-slot-modal.service';
+import { CalendariusSpaceService } from './calendarius-space.service';
+
+@NgModule({
+  providers: [CalendariusSpaceService, HappeningSlotModalService],
+})
+export class CalendariusServicesModule {}